round floats in vcf
