+++ haplokit/io/vcf/__init__.py
@@ -2,6 +2,7 @@ from haplokit.io.vcf import headermeta
 from haplokit.io.vcf import infofields
 from haplokit.io.vcf import formatfields
 from haplokit.io.vcf import filters
+from haplokit.io.vcf import util
 from .classes import *
 from .genotypes import Genotype, HaplotypeAlleleLabeler, call_phenotype
 from .contigs import ContigHeader, contig_headers
